Reintroduce filename admin validation logic  This is to catch basic filename validation errors before the uploaded document is sent over to the document download api for further checks.  Also rolling back logic to upload the mime type returned by the document download api with the file to S3.
