+++ app/main/views/template_email_files.py
@@ -146,7 +146,6 @@ def upload_template_email_files(template_id, service_id):
             filename=form.file.data.filename,
             file_contents=form.file.data,
             template_id=template.id,
-            mimetype=form.file.mimetype,
         )
 
         return redirect(
