+++ app/s3_client/s3_template_email_file_upload_client.py
@@ -3,7 +3,7 @@ from notifications_utils.s3 import s3download as utils_s3download
 from notifications_utils.s3 import s3upload as utils_s3upload
 
 
-def upload_template_email_file_to_s3(data, file_location, mimetype):
+def upload_template_email_file_to_s3(data, file_location):
     metadata = {}
 
     utils_s3upload(
@@ -12,7 +12,6 @@ def upload_template_email_file_to_s3(data, file_location, mimetype):
         bucket_name=current_app.config["S3_BUCKET_TEMPLATE_EMAIL_FILES"],
         file_location=file_location,
         metadata=metadata,
-        content_type=mimetype,
     )
