+++ tests/app/main/views/test_template_email_files.py
@@ -1,3 +1,4 @@
+import io
 import uuid
 from io import BytesIO
 from unittest.mock import ANY, Mock, call
@@ -8,6 +9,7 @@ from notifications_python_client.errors import HTTPError
 from notifications_utils.testing.comparisons import AnyInstanceOf, AnyStringMatching
 from werkzeug.datastructures import FileStorage
 
+from app.notify_client.document_download_api_client import DocumentDownloadError
 from tests import UUID4_REGEX_PATTERN
 from tests.conftest import (
     SERVICE_ONE_ID,
@@ -595,7 +597,6 @@ def test_create_file_redirects_to_manage_files_page(
             filename="tests/test_pdf_files/one_page_pdf.pdf",
             file_contents=AnyInstanceOf(FileStorage),
             template_id=fake_uuid,
-            mimetype="application/pdf",
         ),
     ]
     assert normalize_spaces(page.select_one("form .govuk-button")) == "Add to template"
@@ -879,16 +880,16 @@ def test_upload_file_page_requires_file(
         ("tests/spreadsheet_files/equivalents/excel 2007.xlsx", None),
         (
             "tests/spreadsheet_files/equivalents/EXCEL_95.XLS",
-            "Unsupported file type '.xls'. Supported types are: '.csv', '.doc', '.docx', '.jpeg', '.jpg', '.json', '.odt', '.pdf', '.png', '.rtf', '.txt', '.xlsx'",  # noqa: E501
+            ".XLS is not an allowed file format",
         ),
         ("tests/test_img_files/small-but-perfectly-formed.png", None),
         ("tests/test_pdf_files/big.pdf", "The file must be smaller than 2MB"),
         ("tests/text_files/without brackets.txt", None),
         ("tests/text_files/with (brackets).txt", "File name cannot contain brackets"),
-        ("tests/text_files/no extension", "`filename` must end with a file extension. For example, filename.csv"),
+        ("tests/text_files/no extension", "Not an allowed file format"),
     ),
 )
-def test_upload_file_page_validates_extentions(
+def test_upload_file_page_validates_extensions(
     client_request,
     fake_uuid,
     service_one,
@@ -982,7 +983,6 @@ def test_upload_file_does_not_update_template_content(
             bucket_name="test-template-email-files",
             file_location=AnyStringMatching(rf"{SERVICE_ONE_ID}/{UUID4_REGEX_PATTERN}"),
             metadata={},
-            content_type="application/pdf",
         ),
     ]
     assert mock_post.call_args_list == [
@@ -1159,40 +1159,39 @@ def test_upload_file_returns_error_if_file_fails_antivirus_check(
     mocker,
 ):
     service_one["contact_link"] = "https://example.com"
-    filename = "tests/test_pdf_files/simulate_antivirus_scan_failure.pdf"
+    file_content = b"%PDF-1.4 seriously very very very bad malware content"
+    file_object = io.BytesIO(file_content)
+    filename = "simulate_antivirus_scan_failure.pdf"
     mocker.patch(
         "app.service_api_client.get_service_template",
         return_value={
             "data": create_template(
                 template_id=fake_uuid,
                 template_type="email",
-                email_files=[
-                    {
-                        "id": fake_uuid,
-                        "filename": filename,
-                        "link_text": None,
-                        "retention_period": 50,
-                        "validate_users_email": False,
-                    },
-                ],
+                email_files=[],
             )
         },
     )
 
+    mock_file_check_and_antivirus_scan = mocker.patch(
+        "app.document_download_api_client.file_check_and_antivirus_scan",
+        side_effect=DocumentDownloadError(message="File did not pass the virus scan", status_code=400),
+    )
     mock_s3 = mocker.patch("app.s3_client.s3_template_email_file_upload_client.utils_s3upload")
     mock_post = mocker.patch("app.template_email_file_client.post")
     mock_template_update = mocker.patch("app.service_api_client.update_service_template")
-    with open(filename, "rb") as file:
-        page = client_request.post(
-            "main.upload_template_email_files",
-            service_id=SERVICE_ONE_ID,
-            template_id=fake_uuid,
-            _data={"file": file},
-            _expected_status=200,
-        )
-    assert normalize_spaces(page.select_one(".govuk-error-message").text) == (
-        "File did not pass the virus scan"
+
+    page = client_request.post(
+        "main.upload_template_email_files",
+        service_id=SERVICE_ONE_ID,
+        template_id=fake_uuid,
+        _data={"file": (file_object, filename)},
+        _expected_status=200,
     )
+    assert normalize_spaces(page.select_one(".govuk-error-message").text) == ("File did not pass the virus scan")
+    assert mock_file_check_and_antivirus_scan.call_args_list == [
+        call(service_id=service_one["id"], file_name=filename, file_bytes=file_content)
+    ]
     assert mock_template_update.call_args_list == []
     assert mock_s3.call_args_list == []
     assert mock_post.call_args_list == []
