+++ tests/conftest.py
@@ -12,11 +12,12 @@ import html5lib
 import pytest
 from flask import Flask, current_app, url_for
 from notifications_python_client.errors import HTTPError
+from notifications_utils.file_types import EXTENSIONS, is_allowed_file_extension, mime_type_from_extension
 from notifications_utils.url_safe_token import generate_token
 
 from app import create_app, reset_memos, webauthn_server
 from app.constants import REPORT_REQUEST_STORED, LetterLanguageOptions
-from app.notify_client.document_download_api_client import MockDocumentDownloadAPIClient
+from app.notify_client.document_download_api_client import AbstractDocumentDownloadClient, DocumentDownloadError
 
 from . import (
     NotifyBeautifulSoup,
@@ -4665,3 +4666,66 @@ def mock_document_download_api_client(client_request, mocker):
         side_effect=document_download_api_client.file_check_and_antivirus_scan,
     )
     return document_download_api_client
+
+
+class MockDocumentDownloadAPIClient(AbstractDocumentDownloadClient):
+    """
+    This "mock" class is to act as a realistic approximation of the
+    behaviour of the document download api minus the network calls.
+    It is important that any changes in the document download API validation is caught
+    by the admin unit tests.
+    It mimics the exact file name validation logic, errors, and success responses of document download API.
+
+    """
+
+    def __init__(self):
+        self.base_url = "http://mock-document-download-api-url"
+        self.auth_token = "mock-auth-token"
+
+    def file_check_and_antivirus_scan(self, service_id: str, file_name: str, file_bytes: bytes) -> dict:
+        # check file name
+        try:
+            validate_filename(file_name)
+        except ValueError as e:
+            raise DocumentDownloadError(message=str(e), status_code=400) from e
+
+        # simulate failed antivirus scan if the test file name is:
+        # tests/test_pdf_files/simulate_antivirus_scan_failure.pdf
+        if file_name == "simulate_antivirus_scan_failure.pdf":
+            raise DocumentDownloadError(message="File did not pass the virus scan", status_code=400)
+
+        # get the mimetype
+        extension = split_filename(file_name, dotted=False)[1].lower()
+        mimetype = mime_type_from_extension(extension)
+
+        # check file size
+        if len(file_bytes) > (2 * 1024 * 1024):
+            raise DocumentDownloadError(message="The file must be smaller than 2MB", status_code=413)
+
+        return {
+            "mimetype": mimetype,
+        }
+
+
+# TODO These helper functions where copied over from Document Download API and will be moved to notification utils
+def validate_filename(filename):
+    if len(filename) > current_app.config["MAX_CUSTOM_FILENAME_LENGTH"]:
+        raise ValueError(
+            f"`filename` cannot be longer than {current_app.config['MAX_CUSTOM_FILENAME_LENGTH']} characters"
+        )
+
+    if "." not in filename:
+        raise ValueError("`filename` must end with a file extension. For example, filename.csv")
+
+    extension = split_filename(filename, dotted=False)[1].lower()
+    if not is_allowed_file_extension(extension):
+        allowed_file_types = ", ".join(sorted({f"'.{x}'" for x in EXTENSIONS}))
+        raise ValueError(f"Unsupported file type '.{extension}'. Supported types are: {allowed_file_types}")
+
+    return filename
+
+
+def split_filename(filename: str, *, dotted: bool) -> tuple[str, str]:
+    *parts, ext = filename.split(".")
+    name = ".".join(parts)
+    return name, f".{ext}" if dotted else ext
